# Remove unnecessary metadata_expand.max_seq_len_k operations in fa3 to…