**Before: 15us + 18us -> After: 0**

<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

Remove unnecessary `metadata_expand.max_seq_len_k` operations in fa3 - there is no code reference to `metadata_expand.max_seq_len_k` variable anymore

Profile scripts:
```
python3 -m sglang.launch_server --speculative-num-steps 3 --speculative-eagle-topk 2 --speculative-num-draft-tokens 4 --attention-backend fa3 --trust-remote-code --dtype float16 --speculative-algo EAGLE3 --model-path /shared/public/models/meta-llama/Meta-Llama-3.1-8B-Instruct --speculative-draft-model-path /shared/public/elr-models/jamesliu1/sglang-EAGLE3-Llama-3.1-Instruct-8B/e5ed08d66f528a95ce89f5d4fd136a28f6def714 --disable-radix-cache

python3 -m sglang.bench_serving --backend sglang --dataset-name random --random-input 1024 --random-output 512 --request-rate 16 --num-prompt 512 --model-path /shared/public/models/meta-llama/Meta-Llama-3.1-8B-Instruct
```
**Before: 50% of init_forward_metadata_replay_cuda_graph steps calculate `metadata_expand.max_seq_len_k` using `metadata_expand.cache_seqlens_int32.max().item()` (max() & item() function)**
<img width="915" alt="image" src="https://github.com/user-attachments/assets/b5dfb70b-66c7-4983-a0f5-f308b389e932" />

**After: no more max() or item() calculation (15us + 18us=33us) within init_forward_metadata_replay_cuda_graph for `metadata_expand.max_seq_len_k`**
<img width="657" alt="image" src="https://github.com/user-attachments/assets/8ce12d14-4aaf-4b12-bd63-122cc96034d3" />


Validated correctness via gsm8k:
```
python -m sglang.launch_server --model-path /shared/public/models/meta-llama/Meta-Llama-3.1-8B-Instruct --attention-backend fa3

python benchmark/gsm8k/bench_sglang.py --num-shots 8 --num-questions 1319 --parallel 1319 --
data-path /shared/public/data/gsm8k/test.jsonl 

Accuracy: 0.790
Invalid: 0.001
Latency: 16.377 s
Output throughput: 8214.036 token/s
```
## Modifications

<!-- Describe the changes made in this PR. -->

## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.